requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0
django-environ>=0.11,<1.0
Pillow>=10.0,<12.0
whitenoise>=6.7,<7.0
gunicorn>=22.0,<23.0
uvicorn>=0.30,<1.0
plotly>=5.24,<6.0
